fix: reorganize distributing_computing
